requirements.txt
django == 1.11
djangorestframework
django-oauth-toolkit
django-cors-middleware
psycopg2